+++ CHANGELOG.md
@@ -8,6 +8,10 @@
 
 ### Bugfixes
 
+- `attribute-allowed-values` now normalizes boolean attributes before
+  validating, i.e. it will accept `disabled`, `disabled=""` or
+  `disabled="disabled"`. Fixes #13.
+- `input` learned `required` attribute.
 - `querySelector` properly handles attribute selectors with dashes and
   digits. Fixes #15.